Transfer parsing logic to metadata parsing
